REMOVE EXCEPTION OLD SCHOOL IF

+++ src/database/update_thread.py
@@ -66,25 +66,21 @@ class UpdateThread:
 
     def add_tags_to_thread(self, thread_id, tags):
         for tag in tags:
-            try:
-                tag = self.read_queries.tag_lookup(tag.split()[0])
+            tag = self.read_queries.tag_lookup(tag.split()[0])
+            if tag is not None:
                 db.session.add(TagLine(thread_id=thread_id, tag=tag))
                 tag_in_table = Tag.query.filter(Tag.id==tag).first()
                 tag_in_table.count += 1
-                db.session.commit()
-            except(IntegrityError):
-                continue
+            db.session.commit()
 
     def remove_tags_from_thread(self, thread_id, tags):
         for tag in tags:
-            try:
-                tag = self.read_queries.tag_lookup(tag.split()[0])
+            tag = self.read_queries.tag_lookup(tag.split()[0])
+            if tag is not None:
                 db.session.delete(TagLine.query.filter(TagLine.thread_id==thread_id, TagLine.tag==tag).first())
                 tag_in_table = Tag.query.filter(Tag.id==tag).first()
                 tag_in_table.count -= 1
-                db.session.commit()
-            except(IntegrityError):
-                continue
+            db.session.commit()
         
     def merge_threads(self, a, b):
         thread_a = self.read_queries.get_thread_by_id(a)
